remove migrations table, use schema files instead
