+++ databases/catdat/schema/007_special_objects.sql
