+++ databases/catdat/schema/008_special-morphisms.sql
