+++ databases/catdat/schema/010_functor-properties.sql
@@ -12,6 +12,7 @@ CREATE TABLE functor_properties (
 CREATE UNIQUE INDEX functor_properties_lower_id_unique ON functor_properties (lower(id));
 
 CREATE TABLE functor_property_assignments (
+    id INTEGER PRIMARY KEY,
     functor_id TEXT NOT NULL,
     property_id TEXT NOT NULL,
     is_satisfied INTEGER NOT NULL
@@ -19,8 +20,7 @@ CREATE TABLE functor_property_assignments (
     reason TEXT NOT NULL CHECK (length(reason) > 0),
     is_deduced INTEGER NOT NULL DEFAULT FALSE
         CHECK (is_deduced in (TRUE, FALSE)),
-    position INTEGER NOT NULL DEFAULT 0,
-    PRIMARY KEY (functor_id, property_id),
+    UNIQUE (functor_id, property_id),
     FOREIGN KEY (functor_id) REFERENCES functors (id) ON DELETE CASCADE,
     FOREIGN KEY (property_id) REFERENCES functor_properties (id) ON DELETE CASCADE
 );
