+++ package.json
@@ -16,11 +16,11 @@
 		"prepare": "husky",
 		"db:shell": "sqlite3 databases/catdat/catdat.db",
 		"db:kill": "rm -rf databases/catdat/catdat.db",
-		"db:migrate": "tsx databases/catdat/scripts/migrate.ts && tsx databases/app/scripts/migrate.ts",
+		"db:setup": "pnpm db:kill && tsx databases/catdat/scripts/setup.ts && tsx databases/app/scripts/setup.ts",
 		"db:seed": "tsx databases/catdat/scripts/seed.ts",
 		"db:deduce": "tsx databases/catdat/scripts/deduce.ts",
 		"db:test": "tsx databases/catdat/scripts/test.ts",
-		"db:update": "pnpm db:migrate && pnpm db:seed && pnpm db:deduce && pnpm db:test",
+		"db:update": "pnpm db:setup && pnpm db:seed && pnpm db:deduce && pnpm db:test",
 		"db:watch": "tsx databases/catdat/scripts/watch.ts",
 		"db:snapshot": "cp databases/catdat/catdat.db static/databases/catdat-snapshot.db"
 	},
